prepare separateDiscountsAndCondiment function

+++ pos/models/utils.go
@@ -14,7 +14,7 @@ func advanceInvoiceNumber(terminalID int) (string, error) {
 	if err != nil {
 		return "", err
 	}
-	id := (terminal["last_invoice_id"].(int)) + 1
+	id := int(terminal["last_invoice_id"].(float64)) + 1
 	invoiceNumber = fmt.Sprintf("%d-%d", terminal["id"], id)
 	return invoiceNumber, nil
 }
